+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file following [Keep a Changelog](https://keepachangelog.com/) guidelines.
 
+## [Unreleased]
+- Local test pipeline succeeded (2025-10-30 19:20)
+- Local test pipeline succeeded (2025-10-30 20:40)
+
 ## [v0.4.0] - 2025-10-31
 ### Added
 - Automated local testing pipeline using test_local.sh